Squashed commit of the following: TRIVIAL
commit 307f16be97d8f5cf6d7a255d608e12c42d2d9700
Author: Adam Pigg <adam@piggz.co.uk>
Date:   Thu Dec 22 22:52:28 2016 +0000

    Enable barcode plugin after it was accidentally disabled

    Also fix error in property value

+++ src/plugins/barcode/KReportItemBarcode.cpp
@@ -44,7 +44,7 @@ KReportItemBarcode::KReportItemBarcode(const QDomNode & element)
     m_itemValue->setValue(element.toElement().attribute(QLatin1String("report:value")));
     setZ(element.toElement().attribute(QLatin1String("report:z-index")).toDouble());
     m_horizontalAlignment->setValue(element.toElement().attribute(QLatin1String("report:horizontal-align")));
-    m_maxLength->setValue(element.toElement().attribute(QLatin1String("report:barcode-max-length")));
+    m_maxLength->setValue(element.toElement().attribute(QLatin1String("report:barcode-max-length")).toInt());
     m_format->setValue(element.toElement().attribute(QLatin1String("report:barcode-format")));
     parseReportRect(element.toElement());
